requirements.txt
flask
python-dotenv
gunicorn
python-binance